Create variation2
Follow-up on the work we started in class examining and improving the three example program we looked at. Pick your two favorite variations of those programs you came up with and upload them to your GitHub repository named hw3. Be sure to give the programs a meaningful file name when you upload them.